change id type to BigInteger to increase range limit

+++ src/models/user.py
@@ -8,7 +8,7 @@ from db import Base
 class User(Base):
     __tablename__ = "user"
 
-    id = sa.Column(sa.Integer, primary_key=True)
+    id = sa.Column(sa.BigInteger, primary_key=True)
     phone_number = sa.Column(sa.String(20), unique=True, nullable=True)
     first_name = sa.Column(sa.String(256), unique=False, nullable=True)
     last_name = sa.Column(sa.String(256), unique=False, nullable=True)
